PSR-4 autoloading for exceptions
