+++ src/Exception.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * This file is part of the Nextras community extensions of Nette Framework
+ *
+ * @license    New BSD License
+ * @link       https://github.com/nextras/migrations
+ */
+
+namespace Nextras\Migrations;
+
+interface Exception
+{
+}
