+++ src/ExecutionException.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * This file is part of the Nextras community extensions of Nette Framework
+ *
+ * @license    New BSD License
+ * @link       https://github.com/nextras/migrations
+ */
+
+namespace Nextras\Migrations;
+
+/**
+ * Executing migration has failed.
+ */
+class ExecutionException extends RuntimeException
+{
+}
